Smart playlist: year range filter + refactor rules form into shared component
